refactor: files/folders moving and renaming
