+++ src/bot/events/interaction-create/embed/role-grant/create/button.ts
@@ -1,9 +1,9 @@
 import { Events, GuildMember, Interaction } from "discord.js"
-import { Event } from "../../event"
-import { generateCustomId } from "../../../../utils/io"
-import { EVENT_PATH } from "../.."
-import { discordReply, getDiscordBot, getDiscordRole } from "../../../../utils/discord"
-import { log } from "../../../../utils/logger"
+import { Event } from "../../../../event"
+import { generateCustomId } from "../../../../../../utils/io"
+import { EVENT_PATH } from "../../../.."
+import { discordReply, getDiscordBot, getDiscordRole } from "../../../../../../utils/discord"
+import { log } from "../../../../../../utils/logger"
 import { ERR, MSG } from "./messages"
 import { assertMember, assertMemberHasRole, assertRole, assertRoleManageable, getButtonCustomId } from "./validators"
