+++ src/bot/events/interaction-create/embed/role-grant/create/modal.ts
@@ -1,9 +1,9 @@
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle, EmbedBuilder, Events, Interaction, roleMention } from "discord.js"
-import { Event } from "../../event"
-import { COMMAND_EMBED_ID } from "../../../commands/embed/buttonCreate"
-import { discordReply, getDiscordBot, getDiscordChannel, getDiscordRole } from "../../../../utils/discord"
-import { EVENT_EMBED_ID } from "./buttonCreate"
-import { log } from "../../../../utils/logger"
+import { Event } from "../../../../event"
+import { COMMAND_EMBED_ID } from "../../../../../commands/embed/role-grant/create/button"
+import { discordReply, getDiscordBot, getDiscordChannel, getDiscordRole } from "../../../../../../utils/discord"
+import { EVENT_EMBED_ID } from "./button"
+import { log } from "../../../../../../utils/logger"
 import { ERR } from "./messages"
 import { assertBotCanPost, assertRole, assertRoleManageable, assertTextChannel, getModalCustomId } from "./validators"
